requirements.txt
aiohappyeyeballs @ file:///C:/b/abs_e9fm5mk6ak/croot/aiohappyeyeballs_1734469429350/work
aiohttp @ file:///C:/b/abs_d4_gdjffav/croot/aiohttp_1734687146788/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
aiosqlite @ file:///C:/b/abs_9djc_0pyi3/croot/aiosqlite_1683773915844/work
async-timeout @ file:///C:/b/abs_d8yc2t_bxy/croot/async-timeout_1732663677668/work
attrs @ file:///C:/b/abs_89hmquz5ga/croot/attrs_1734533130810/work
click @ file:///C:/b/abs_f9ihnt72pu/croot/click_1698129847492/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
discord.py==2.5.0
frozenlist @ file:///C:/b/abs_06ctmb1zeo/croot/frozenlist_1730903113463/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
multidict @ file:///C:/b/abs_19e3ubo2ew/croot/multidict_1730905504444/work
propcache @ file:///C:/b/abs_d6o8xbonwb/croot/propcache_1732304003668/work
python-dotenv @ file:///C:/b/abs_edyrwjya7k/croot/python-dotenv_1669132572913/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
yarl @ file:///C:/b/abs_281qby2vim/croot/yarl_1732546854547/work
